Moved LegalApproval from operationalplanning app to legalapproval app and also moved respective admin forms around.
